requirements.txt
numpy==2.2.4
Pillow==11.1.0
Requests==2.32.3
streamlit==1.39.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
